+++ rappi4/src/pages/FeedPage/styled.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+
+export const Image = styled.img`
+ width: 328px;
+ height: 120px;
+ margin: 0 0 12px;
+ object-fit: contain;
+`
+
+export const CardRestaurante = styled.div`
+
+  width: 360px;
+  height: 196px;
+  margin: 50px 0 0;
+  padding: 8px 16px 0;
+
+`
+
+
+   
+
