+++ rappi4/src/services/Global/GlobalState.js
@@ -8,34 +8,10 @@ import { BASE_URL } from '../../constants/urls'
 
 const GlobalState = (props) => {
 
-    useEffect(() => {
-        PegarRestaurantes()
-    }, [])
-
+   
 
     const [restaurants, setRestaurantes] = useState([])
-
-
-    const PegarRestaurantes = () => {
-        const url = `${BASE_URL}/restaurants`
-        const token = localStorage.getItem("token")
-        const header = {
-            headers: {
-                auth: token
-            },
-        }
-        axios.get(url, header)
-            .then((resp) => {
-                setRestaurantes(resp.data.restaurants)
-                console.log(resp.data.restaurants)
-            })
-            .catch((err) => {
-                console.log("errooooo", err.response)
-            })
-    }
-    console.log(restaurants)
-
-    const values = { restaurants, setRestaurantes, PegarRestaurantes}
+    const values = { restaurants, setRestaurantes, }
     const Provider = GlobalContext.Provider
 
     return (
